+++ .github/scripts/test_issue255_update_ui_write_suppression_runtime.mjs
@@ -15,8 +15,8 @@ import { instrumentSource } from "../../tools/build-render-probe-userscript.mjs"
 const ROOT = path.resolve(path.dirname(fileURLToPath(import.meta.url)), "../..");
 const SOURCE_PATH = path.join(ROOT, "src/MissionChief_Map_Command_Toolkit.user.js");
 const BASELINE_PATH = path.join(ROOT, "docs/audits/issue-255/unchanged-update-ui.json");
-const EXPECTED_VERSION = "10.3.2";
-const EXPECTED_SHA = "c41839cac588837a6d924bb3362a34f49466596a46d6d0ffb0795092a8c88687";
+const EXPECTED_VERSION = "10.3.3";
+const EXPECTED_SHA = "d8365695ea572d7ba27e3e184873372caf05aebc42a23cb3ac36648f67ad2947";
 const REPEATS = 25;
 const HELPER_NAMES = ["normaliseDiscordReportComplexity", "discordReportComplexityAtLeast", "updateUiToggleClass", "updateUiSetStyleProperty", "updateUiSetAttribute", "updateUiSetDataset", "updateUiSetProperty", "updateUiSetText", "commandInterfaceApplySearch", "updateCommandInterfaceHeader"];
