+++ .github/scripts/test_issue618_command_palette_contract.py
@@ -18,7 +18,7 @@ def main() -> int:
     source = SOURCE.read_text(encoding="utf-8")
     metadata = re.search(r"(?m)^//\s*@version\s+([^\s]+)$", source)
     runtime = re.search(r"version:\s*'([^']+)'", source)
-    assert metadata and runtime and metadata.group(1) == runtime.group(1) == "10.3.2"
+    assert metadata and runtime and metadata.group(1) == runtime.group(1) == "10.3.3"
 
     palette = section(source, "    function commandPaletteNormalise(", "    function closeHelpCenter(")
     for required in [
