+++ .github/scripts/test_toolkit_analytics_contract.py
@@ -15,8 +15,8 @@ def require(token: str) -> None:
 
 
 for token in (
-    "// @version      10.3.2",
-    "version: '10.3.2'",
+    "// @version      10.3.3",
+    "version: '10.3.3'",
     "https://tkb-gaming.scot/api/toolkit-analytics.php",
     "runtimeFetch(TOOLKIT_ANALYTICS_ENDPOINT",
     "globalPrivacyControl",
